+++ packages/moonstone/Scroller/Scroller.js
@@ -16,102 +16,97 @@
  * @exports ScrollerBase
  */
 
-
-import {ScrollerBase as UiScrollerBase} from '@enact/ui/Scroller';
+import {ResizeContext} from '@enact/ui/Resizable';
 import PropTypes from 'prop-types';
-import React, {useEffect} from 'react';
+import React from 'react';
 
-import Scrollable from '../Scrollable';
+import {useScrollableComponentizable} from '../Scrollable';
+import Scrollbar from '../Scrollable/Scrollbar';
+import Skinnable from '../Skinnable';
 
+import {I18nContextDecorator} from '@enact/i18n/I18nDecorator';
+import SpotlightContainerDecorator from '@enact/spotlight/SpotlightContainerDecorator';
+
+import ScrollerBase from './ScrollerBase';
 import useSpottable from './useSpottable';
 
 /**
- * A Moonstone-styled base component for [Scroller]{@link moonstone/Scroller.Scroller}.
- * In most circumstances, you will want to use the
- * [SpotlightContainerDecorator]{@link spotlight/SpotlightContainerDecorator.SpotlightContainerDecorator}
- * and the Scrollable version, [Scroller]{@link moonstone/Scroller.Scroller}.
+ * A Moonstone-styled Scroller, Scrollable applied.
  *
- * @function ScrollerBase
+ * Usage:
+ * ```
+ * <Scroller>Scroll me.</Scroller>
+ * ```
+ *
+ * @class Scroller
  * @memberof moonstone/Scroller
- * @extends ui/Scroller.ScrollerBase
+ * @extends moonstone/Scroller.ScrollerBase
  * @ui
  * @public
  */
-let ScrollerBase = (props) => {
-	/*
-	 * Dependencies
-	 */
+const ScrollerComp = (props) => {
+	const {
+		childWrapper: ChildWrapper,
+		isHorizontalScrollbarVisible,
+		isVerticalScrollbarVisible,
 
-	const {uiChildAdapter} = props;
-
-	/*
-	 * Hooks
-	 */
-
-	const {calculatePositionOnFocus, focusOnNode, setContainerDisabled} = useSpottable(props, {uiChildAdapter});
-
-	useEffect(() => {
-		props.setChildAdapter({
-			calculatePositionOnFocus,
-			focusOnNode,
-			setContainerDisabled
-		});
-	}, []);
-
-	/*
-	 * Render
-	 */
-
-	const propsObject = Object.assign({}, props);
-
-	delete propsObject.dangerouslyContainsInScrollable;
-	delete propsObject.onUpdate;
-	delete propsObject.scrollAndFocusScrollbarButton;
-	delete propsObject.setChildAdapter;
-	delete propsObject.spotlightId;
-	delete propsObject.uiScrollableAdapter;
+		resizeContextProps,
+		scrollableContainerProps,
+		flexLayoutProps,
+		childWrapperProps,
+		childProps,
+		verticalScrollbarProps,
+		horizontalScrollbarProps
+	} = useScrollableComponentizable(props);
 
 	return (
-		<UiScrollerBase {...propsObject} />
+		<ResizeContext.Provider {...resizeContextProps}>
+			<div {...scrollableContainerProps}>
+				<div {...flexLayoutProps}>
+					<ChildWrapper {...childWrapperProps}>
+						<ScrollerBase {...childProps} />
+					</ChildWrapper>
+					{isVerticalScrollbarVisible ? <Scrollbar {...verticalScrollbarProps} /> : null}
+				</div>
+				{isHorizontalScrollbarVisible ? <Scrollbar {...horizontalScrollbarProps} /> : null}
+			</div>
+		</ResizeContext.Provider>
 	);
 };
 
-ScrollerBase.displayName = 'ScrollerBase';
-
-ScrollerBase.propTypes = /** @lends moonstone/Scroller.ScrollerBase.prototype */ {
-	/**
-	 * Called when [Scroller]{@link moonstone/Scroller.Scroller} updates.
-	 *
-	 * @type {function}
-	 * @private
-	 */
-	onUpdate: PropTypes.func,
-
-	/**
-	 * `true` if rtl, `false` if ltr.
-	 *
-	 * @type {Boolean}
-	 * @private
-	 */
-	rtl: PropTypes.bool,
-
-	/**
-	 * Called when [Scroller]{@link moonstone/Scroller.Scroller} should be scrolled
-	 * and the focus should be moved to a scrollbar button.
-	 *
-	 * @type {function}
-	 * @private
-	 */
-	scrollAndFocusScrollbarButton: PropTypes.func,
-
-	/**
-	 * The spotlight id for the component.
-	 *
-	 * @type {String}
-	 * @private
-	 */
-	spotlightId: PropTypes.string
+ScrollerComp.propTypes = /** @lends moonstone/Scroller.Scroller.prototype */ {
+	direction: PropTypes.oneOf(['both', 'horizontal', 'vertical'])
 };
+
+ScrollerComp.defaultProps = {
+	direction: 'both',
+
+	'data-spotlight-container-disabled': false,
+	focusableScrollbar: false,
+	overscrollEffectOn: {
+		arrowKey: false,
+		drag: false,
+		pageKey: false,
+		scrollbarButton: false,
+		wheel: true
+	},
+	preventBubblingOnKeyDown: 'none',
+	type: 'JS'
+};
+
+const Scroller = Skinnable(
+	SpotlightContainerDecorator(
+		{
+			overflow: true,
+			preserveId: true,
+			restrict: 'self-first'
+		},
+		I18nContextDecorator(
+			{rtlProp: 'rtl'},
+			ScrollerComp
+		)
+	)
+);
 
 /**
  * Allows 5-way navigation to the scrollbar controls. By default, 5-way will
@@ -176,39 +171,6 @@ ScrollerBase.propTypes = /** @lends moonstone/Scroller.ScrollerBase.prototype */
  * @default $L('scroll up')
  * @public
  */
-
-/**
- * A Moonstone-styled Scroller, Scrollable applied.
- *
- * Usage:
- * ```
- * <Scroller>Scroll me.</Scroller>
- * ```
- *
- * @class Scroller
- * @memberof moonstone/Scroller
- * @extends moonstone/Scroller.ScrollerBase
- * @ui
- * @public
- */
-const Scroller = (props) => {
-	return (
-		<Scrollable
-			{...props}
-			childRenderer={(scrollerProps) => { // eslint-disable-line react/jsx-no-bind
-				return <ScrollerBase {...scrollerProps} />;
-			}}
-		/>
-	);
-};
-
-Scroller.propTypes = /** @lends moonstone/Scroller.Scroller.prototype */ {
-	direction: PropTypes.oneOf(['both', 'horizontal', 'vertical'])
-};
-
-Scroller.defaultProps = {
-	direction: 'both'
-};
 
 export default Scroller;
 export {
