+++ packages/moonstone/VirtualList/VirtualList.js
@@ -8,48 +8,21 @@
  */
 
 import kind from '@enact/core/kind';
+import {ResizeContext} from '@enact/ui/Resizable';
 import {gridListItemSizeShape, itemSizesShape} from '@enact/ui/VirtualList';
 import PropTypes from 'prop-types';
 import React from 'react';
 import warning from 'warning';
 
-import Scrollable from '../Scrollable';
+import {useScrollableComponentizable} from '../Scrollable';
+import Scrollbar from '../Scrollable/Scrollbar';
+import Skinnable from '../Skinnable';
+
+import {I18nContextDecorator} from '@enact/i18n/I18nDecorator';
+import SpotlightContainerDecorator from '@enact/spotlight/SpotlightContainerDecorator';
+
 import VirtualListBase from './VirtualListBase';
-
-/* eslint-disable enact/prop-types */
-const listItemsRenderer = (props) => {
-	const {
-		cc,
-		handlePlaceholderFocus,
-		itemContainerRef: initUiItemContainerRef,
-		needsScrollingPlaceholder,
-		primary,
-		role,
-		SpotlightPlaceholder
-	} = props;
-
-	return (
-		<React.Fragment>
-			{cc.length ? (
-				<div ref={initUiItemContainerRef} role={role}>{cc}</div>
-			) : null}
-			{primary ? null : (
-				<SpotlightPlaceholder
-					data-index={0}
-					data-vl-placeholder
-					// a zero width/height element can't be focused by spotlight so we're giving
-					// the placeholder a small size to ensure it is navigable
-					onFocus={handlePlaceholderFocus}
-					style={{width: 10}}
-				/>
-			)}
-			{needsScrollingPlaceholder ? (
-				<SpotlightPlaceholder />
-			) : null}
-		</React.Fragment>
-	);
-};
-/* eslint-enable enact/prop-types */
+import virtualListItemsRenderer from './virtualListItemsRenderer';
 
 /**
  * A Moonstone-styled scrollable and spottable virtual list component.
@@ -60,74 +33,116 @@ const listItemsRenderer = (props) => {
  * @ui
  * @public
  */
-const VirtualList = kind({
-	name: 'VirtualList',
+const VirtualListComp = ({itemSize, role, ...rest}) => {
+	const props = itemSize && itemSize.minSize ?
+		{
+			itemSize: itemSize.minSize,
+			itemSizes: itemSize.size
+		} :
+		{
+			itemSize
+		};
 
-	propTypes: /** @lends moonstone/VirtualList.VirtualList.prototype */ {
-		/**
-		 * Size of an item for the VirtualList; valid value is a number generally.
-		 * For different item size, value is an object that has `minSize`
-		 * and `size` as properties.
-		 * If the direction for the list is vertical, itemSize means the height of an item.
-		 * For horizontal, it means the width of an item.
-		 *
-		 * Usage:
-		 * ```
-		 * <VirtualList itemSize={ri.scale(72)} />
-		 * ```
-		 *
-		 * @type {Number|ui/VirtualList.itemSizesShape}
-		 * @required
-		 * @public
-		 */
-		itemSize: PropTypes.oneOfType([PropTypes.number, itemSizesShape]).isRequired,
+	warning(
+		!rest.itemSizes || !rest.cbScrollTo,
+		'VirtualList with `minSize` in `itemSize` prop does not support `cbScrollTo` prop'
+	);
 
-		cbScrollTo: PropTypes.func,
-		direction: PropTypes.oneOf(['horizontal', 'vertical']),
-		focusableScrollbar: PropTypes.bool,
-		itemSizes: PropTypes.array,
-		preventBubblingOnKeyDown: PropTypes.oneOf(['none', 'programmatic']),
-		role: PropTypes.string
+	const {
+		childWrapper: ChildWrapper,
+		isHorizontalScrollbarVisible,
+		isVerticalScrollbarVisible,
+
+		resizeContextProps,
+		scrollableContainerProps,
+		flexLayoutProps,
+		childWrapperProps,
+		childProps,
+		verticalScrollbarProps,
+		horizontalScrollbarProps
+	} = useScrollableComponentizable({...rest, ...props});
+
+	return (
+		<ResizeContext.Provider {...resizeContextProps}>
+			<div {...scrollableContainerProps}>
+				<div {...flexLayoutProps}>
+					<ChildWrapper {...childWrapperProps}>
+						<VirtualListBase
+							{...childProps}
+							focusableScrollbar={rest.focusableScrollbar}
+							itemsRenderer={virtualListItemsRenderer}
+							role={role}
+						/>
+					</ChildWrapper>
+					{isVerticalScrollbarVisible ? <Scrollbar {...verticalScrollbarProps} /> : null}
+				</div>
+				{isHorizontalScrollbarVisible ? <Scrollbar {...horizontalScrollbarProps} /> : null}
+			</div>
+		</ResizeContext.Provider>
+	);
+};
+
+VirtualListComp.displayName = 'VirtualList';
+
+VirtualListComp.propTypes = /** @lends moonstone/VirtualList.VirtualList.prototype */ {
+	/**
+	 * Size of an item for the VirtualList; valid value is a number generally.
+	 * For different item size, value is an object that has `minSize`
+	 * and `size` as properties.
+	 * If the direction for the list is vertical, itemSize means the height of an item.
+	 * For horizontal, it means the width of an item.
+	 *
+	 * Usage:
+	 * ```
+	 * <VirtualList itemSize={ri.scale(72)} />
+	 * ```
+	 *
+	 * @type {Number|ui/VirtualList.itemSizesShape}
+	 * @required
+	 * @public
+	 */
+	itemSize: PropTypes.oneOfType([PropTypes.number, itemSizesShape]).isRequired,
+
+	cbScrollTo: PropTypes.func,
+	direction: PropTypes.oneOf(['horizontal', 'vertical']),
+	focusableScrollbar: PropTypes.bool,
+	itemSizes: PropTypes.array,
+	preventBubblingOnKeyDown: PropTypes.oneOf(['none', 'programmatic']),
+	role: PropTypes.string
+};
+
+VirtualListComp.defaultProps = {
+	direction: 'vertical',
+	focusableScrollbar: false,
+	preventBubblingOnKeyDown: 'programmatic',
+	role: 'list',
+
+	'data-spotlight-container-disabled': false,
+	focusableScrollbar: false,
+	overscrollEffectOn: {
+		arrowKey: false,
+		drag: false,
+		pageKey: false,
+		scrollbarButton: false,
+		wheel: true
 	},
+	preventBubblingOnKeyDown: 'none',
+	type: 'JS'
+};
 
-	defaultProps: {
-		direction: 'vertical',
-		focusableScrollbar: false,
-		preventBubblingOnKeyDown: 'programmatic',
-		role: 'list'
-	},
-
-	render: ({itemSize, role, ...rest}) => {
-		const props = itemSize && itemSize.minSize ?
-			{
-				itemSize: itemSize.minSize,
-				itemSizes: itemSize.size
-			} :
-			{
-				itemSize
-			};
-
-		warning(
-			!rest.itemSizes || !rest.cbScrollTo,
-			'VirtualList with `minSize` in `itemSize` prop does not support `cbScrollTo` prop'
-		);
-
-		return (
-			<Scrollable
-				{...rest}
-				{...props}
-				childRenderer={(childProps) => ( // eslint-disable-line react/jsx-no-bind
-					<VirtualListBase
-						{...childProps}
-						focusableScrollbar={rest.focusableScrollbar}
-						itemsRenderer={listItemsRenderer}
-						role={role}
-					/>
-				)}
-			/>
-		);
-	}
-});
+const VirtualList = Skinnable(
+	SpotlightContainerDecorator(
+		{
+			overflow: true,
+			preserveId: true,
+			restrict: 'self-first'
+		},
+		I18nContextDecorator(
+			{rtlProp: 'rtl'},
+			VirtualListComp
+		)
+	)
+);
 
 /**
  * A Moonstone-styled scrollable and spottable virtual grid list component.
@@ -138,61 +153,104 @@ const VirtualList = kind({
  * @ui
  * @public
  */
-const VirtualGridList = kind({
-	name: 'VirtualGridList',
+const VirtualGridListComp = ({role, ...rest}) => {
+	const {
+		childWrapper: ChildWrapper,
+		isHorizontalScrollbarVisible,
+		isVerticalScrollbarVisible,
 
-	propTypes: /** @lends moonstone/VirtualList.VirtualGridList.prototype */ {
-		/**
-		 * Size of an item for the VirtualGridList; valid value is an object that has `minWidth`
-		 * and `minHeight` as properties.
-		 *
-		 * Usage:
-		 * ```
-		 * <VirtualGridList
-		 * 	itemSize={{
-		 * 		minWidth: ri.scale(180),
-		 * 		minHeight: ri.scale(270)
-		 * 	}}
-		 * />
-		 * ```
-		 *
-		 * @type {ui/VirtualList.gridListItemSizeShape}
-		 * @required
-		 * @public
-		 */
-		itemSize: gridListItemSizeShape.isRequired,
+		resizeContextProps,
+		scrollableContainerProps,
+		flexLayoutProps,
+		childWrapperProps,
+		childProps,
+		verticalScrollbarProps,
+		horizontalScrollbarProps
+	} = useScrollableComponentizable(rest);
 
-		cbScrollTo: PropTypes.func,
-		direction: PropTypes.oneOf(['horizontal', 'vertical']),
-		focusableScrollbar: PropTypes.bool,
-		itemSizes: PropTypes.array,
-		preventBubblingOnKeyDown: PropTypes.oneOf(['none', 'programmatic']),
-		role: PropTypes.string
+	return (
+		<ResizeContext.Provider {...resizeContextProps}>
+			<div {...scrollableContainerProps}>
+				<div {...flexLayoutProps}>
+					<ChildWrapper {...childWrapperProps}>
+						<VirtualListBase
+							{...childProps}
+							focusableScrollbar={rest.focusableScrollbar}
+							itemsRenderer={virtualListItemsRenderer}
+							role={role}
+						/>
+					</ChildWrapper>
+					{isVerticalScrollbarVisible ? <Scrollbar {...verticalScrollbarProps} /> : null}
+				</div>
+				{isHorizontalScrollbarVisible ? <Scrollbar {...horizontalScrollbarProps} /> : null}
+			</div>
+		</ResizeContext.Provider>
+	);
+};
+
+VirtualGridListComp.displayName = 'VirtualGridList';
+
+VirtualGridListComp.propTypes = /** @lends moonstone/VirtualList.VirtualGridList.prototype */ {
+	/**
+	 * Size of an item for the VirtualGridList; valid value is an object that has `minWidth`
+	 * and `minHeight` as properties.
+	 *
+	 * Usage:
+	 * ```
+	 * <VirtualGridList
+	 * 	itemSize={{
+	 * 		minWidth: ri.scale(180),
+	 * 		minHeight: ri.scale(270)
+	 * 	}}
+	 * />
+	 * ```
+	 *
+	 * @type {ui/VirtualList.gridListItemSizeShape}
+	 * @required
+	 * @public
+	 */
+	itemSize: gridListItemSizeShape.isRequired,
+
+	cbScrollTo: PropTypes.func,
+	direction: PropTypes.oneOf(['horizontal', 'vertical']),
+	focusableScrollbar: PropTypes.bool,
+	itemSizes: PropTypes.array,
+	preventBubblingOnKeyDown: PropTypes.oneOf(['none', 'programmatic']),
+	role: PropTypes.string
+};
+
+VirtualGridListComp.defaultProps = {
+	direction: 'vertical',
+	focusableScrollbar: false,
+	preventBubblingOnKeyDown: 'programmatic',
+	role: 'list',
+
+	'data-spotlight-container-disabled': false,
+	focusableScrollbar: false,
+	overscrollEffectOn: {
+		arrowKey: false,
+		drag: false,
+		pageKey: false,
+		scrollbarButton: false,
+		wheel: true
 	},
+	preventBubblingOnKeyDown: 'none',
+	type: 'JS'
+};
 
-	defaultProps: {
-		direction: 'vertical',
-		focusableScrollbar: false,
-		preventBubblingOnKeyDown: 'programmatic',
-		role: 'list'
-	},
-
-	render: ({role, ...rest}) => {
-		return (
-			<Scrollable
-				{...rest}
-				childRenderer={(childProps) => ( // eslint-disable-line react/jsx-no-bind
-					<VirtualListBase
-						{...childProps}
-						focusableScrollbar={rest.focusableScrollbar}
-						itemsRenderer={listItemsRenderer}
-						role={role}
-					/>
-				)}
-			/>
-		);
-	}
-});
+const VirtualGridList = Skinnable(
+	SpotlightContainerDecorator(
+		{
+			overflow: true,
+			preserveId: true,
+			restrict: 'self-first'
+		},
+		I18nContextDecorator(
+			{rtlProp: 'rtl'},
+			VirtualGridListComp
+		)
+	)
+);
 
 export default VirtualList;
 export {
